\name{coGPS-package}
\alias{coGPS-package}
\alias{coGPS}
\docType{package}
\title{Cancer Outlier Gene Profile Sets}

\description{Gene Set Enrichment Analysis of P-value based statistics for outlier gene detection in dataset merged from multiple studies}

\author{
Yingying Wei, Michael Ochs
Maintainer: Yingying Wei <ywei@jhsph.edu>
}
\references{
Wei, Y., Hennessey, P., Gaykalova, D., Califano, J.A., Ochs, M.F., (2011) Cancer Outlier Gene Profile Sets Elucidate Pathways in Head and Neck Squamous Cell Carcinoma.
}

\keyword{Microarray, Bioinformatics, DifferentialExpression}
